fixing minor issue in litmusctl
